move date logic to mai n, bind shortcut listener to input
